worker: more tests
